fix code-style and other minor fixes

+++ sim/sim_generator.c
@@ -10,13 +10,13 @@ static void generator_emit(GENERATOR *gen)
      int             intensity;
      long long       timestamp_ms;
      struct timespec ts;
- 
+
      sensor_id = rand() % gen->sensor_count;
      intensity = MIN_INTENSITY + rand() % (MAX_INTENSITY - MIN_INTENSITY + 1);
- 
+
      clock_gettime(CLOCK_REALTIME, &ts);
      timestamp_ms = ts.tv_sec * 1000LL + ts.tv_nsec / 1000000LL;
- 
+
      printf("%d %d %lld\n", sensor_id, intensity, timestamp_ms);
      fflush(stdout);
 }
@@ -58,9 +58,9 @@ void generator_tick(GENERATOR *gen)
      double          roll;
      int             i;
      struct timespec ts;
- 
+
      roll = (double)rand() / RAND_MAX;
- 
+
      if (roll < gen->burst_prob)
      {
           for (i = 0; i < gen->burst_size; i++)
@@ -72,7 +72,7 @@ void generator_tick(GENERATOR *gen)
      {
           generator_emit(gen);
           ts.tv_sec  = 0;
-          ts.tv_nsec = (long long)(1000000000.0 / gen->frequency);
+          ts.tv_nsec = (long)(1000000000.0 / gen->frequency);
           nanosleep(&ts, NULL);
      }
 }
